# api/openapi.yaml
openapi: 3.0.3
info:
  title: AgentGuard Policy Engine API
  description: |
    **AgentGuard** is a runtime security platform for AI agents. It provides policy-based evaluation,
    audit trails, PII detection, MCP server governance, agent hierarchy management, and OWASP compliance reporting.

    ## Authentication

    AgentGuard uses three authentication schemes:

    - **API Key** (`x-api-key` header): Your tenant API key, prefixed `ag_live_`. Used for most authenticated endpoints.
    - **Agent Key** (`x-api-key` header): Agent-scoped key, prefixed `ag_agent_`. Used for evaluate calls from specific agents.
    - **Admin Key** (`x-admin-key` header or `x-api-key`): Platform admin key. Required for admin-only endpoints.

    ## Rate Limits

    - Unauthenticated: 10 requests/min per IP
    - Authenticated: 100 requests/min per IP
    - Signup: 5 per hour per IP

    ## Base URL

    Production: `https://api.agentguard.tech/api/v1`
  version: 0.10.0
  contact:
    name: AgentGuard Support
    url: https://agentguard.tech
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.agentguard.tech/api/v1
    description: Production
  - url: http://localhost:3000/api/v1
    description: Local development
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Tenant API key (prefixed `ag_live_`) or agent key (prefixed `ag_agent_`)
    AdminKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Platform admin key (set via ADMIN_KEY environment variable)
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT bearer token (future use)
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
          example: "Invalid API key"
        details:
          type: array
          items:
            type: string
          description: Additional error details (validation errors)
    PolicyDecision:
      type: object
      required: [result, riskScore, reason]
      properties:
        result:
          type: string
          enum: [allow, block, monitor, require_approval]
          description: Policy decision outcome
        riskScore:
          type: integer
          minimum: 0
          maximum: 1000
          description: Computed risk score (0=safe, 1000=maximum risk)
        reason:
          type: string
          description: Human-readable explanation of the decision
        matchedRuleId:
          type: string
          nullable: true
          description: ID of the policy rule that matched
        durationMs:
          type: number
          description: Evaluation duration in milliseconds
        approvalId:
          type: string
          nullable: true
          description: Approval request ID (only when result=require_approval)
        piiDetected:
          type: boolean
          description: Whether PII was detected in the request params
        piiEntityCount:
          type: integer
          description: Number of PII entities detected
...
    PolicyRule:
      type: object
      required: [id, action, when]
      properties:
        id:
          type: string
        description:
          type: string
        priority:
          type: integer
          minimum: 0
          maximum: 1000
        action:
          type: string
          enum: [allow, block, monitor, require_approval]
        when:
          type: array
          items:
            type: object
        severity:
          type: string
          enum: [low, medium, high, critical]
        tags:
          type: array
          items:
            type: string
        riskBoost:
          type: integer
    PolicyDocument:
      type: object
      required: [id, name, version, default, rules]
      properties:
        id:
          type: string
        name:
          type: string
        version:
          type: string
        default:
          type: string
          enum: [allow, block, monitor]
        rules:
          type: array
          items:
            $ref: '#/components/schemas/PolicyRule'
...
paths:
...
  /evaluate:
    get:
      tags: [Evaluate]
      summary: Evaluate endpoint info
      description: Returns a description of the evaluate endpoint and usage instructions. No authentication required.
      operationId: getEvaluateInfo
      security: []
      responses:
        '200':
          description: Endpoint info
          content:
            application/json:
              schema:
                type: object
                properties:
                  endpoint:
                    type: string
                  method:
                    type: string
                  description:
                    type: string
                  requiredFields:
                    type: array
                    items:
                      type: string
                  example:
                    type: object
    post:
      tags: [Evaluate]
      summary: Evaluate an agent action
      description: |
        The core evaluation endpoint. Submits an agent action (tool call) to the policy engine
        and receives a decision: `allow`, `block`, `monitor`, or `require_approval`.

        Authentication is optional — unauthenticated calls use the default policy.
        Authenticated calls use tenant-specific custom policies.
      operationId: evaluate
      security:
        - ApiKeyAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tool]
              properties:
                tool:
                  type: string
                  maxLength: 200
                  pattern: '^[a-zA-Z0-9_.\-:]+$'
                  description: Tool name being requested
                  example: "file_read"
                params:
                  type: object
                  description: Tool parameters (key-value pairs)
                  example:
                    path: "/etc/config.yaml"
                agentId:
                  type: string
                  description: Agent identifier for scoped evaluation
                sessionId:
                  type: string
                  description: Session identifier for grouping audit events
                context:
                  type: object
                  properties:
                    userId:
                      type: string
                    environment:
                      type: string
                      example: "production"
                messageHistory:
                  type: array
                  description: Recent conversation history for context-aware detection
                  items:
                    type: object
                    required: [role, content]
                    properties:
                      role:
                        type: string
                        enum: [user, assistant, system]
                      content:
                        type: string
            examples:
              allow_example:
                summary: File read (typically allowed)
                value:
                  tool: "file_read"
                  params:
                    path: "/var/log/app.log"
                  agentId: "my-agent-001"
              block_example:
                summary: Shell exec (typically blocked)
                value:
                  tool: "shell_exec"
                  params:
                    command: "rm -rf /tmp"
                  agentId: "my-agent-001"
              approval_example:
                summary: Sensitive DB query (requires approval)
                value:
                  tool: "db_query"
                  params:
                    table: "users"
                  agentId: "my-agent-001"
      responses:
        '200':
          description: Policy decision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyDecision'
              examples:
                allow:
                  value:
                    result: allow
                    riskScore: 10
                    reason: "Matched rule: allow-read-operations"
                    matchedRuleId: "rule-allow-read"
                    durationMs: 3.2
                    approvalId: null
                block:
                  value:
                    result: block
                    riskScore: 500
                    reason: "Shell execution is blocked by security policy"
                    matchedRuleId: "rule-block-shell"
                    durationMs: 2.8
                    approvalId: null
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /policy:
    get:
      tags: [Policy]
      summary: Get active policy
      description: Retrieve the tenant's current effective policy (custom or default).
      operationId: getPolicy
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Current policy
          content:
            application/json:
              schema:
                type: object
                properties:
                  tenantId:
                    type: string
                  isCustom:
                    type: boolean
                  policy:
                    $ref: '#/components/schemas/PolicyDocument'

    put:
      tags: [Policy]
      summary: Update policy
      description: |
        Replace the tenant's custom policy with a new set of rules.
        Accepts either a full PolicyDocument or an array of PolicyRule objects.
        Maximum 500 rules per policy.
      operationId: updatePolicy
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: array
                  items:
                    $ref: '#/components/schemas/PolicyRule'
                  description: Array of policy rules
                - $ref: '#/components/schemas/PolicyDocument'
            example:
              rules:
                - id: "my-rule-001"
                  description: "Block file deletion"
                  priority: 100
                  action: "block"
                  when:
                    - tool:
                        in: ["file_delete", "rm"]
                  severity: "high"
      responses:
        '200':
          description: Policy updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  tenantId:
                    type: string
                  ruleCount:
                    type: integer
                  message:
                    type: string
                  policy:
                    $ref: '#/components/schemas/PolicyDocument'
        '400':
          description: Invalid policy rules
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  details:
                    type: array
                    items:
                      type: string

  # ── APPROVALS ────────────────────────────────────────────────────────────
...
  /policy/bundle:
    get:
      tags: [Policy]
      summary: Get compiled policy bundle
      description: |
        Returns the compiled `PolicyBundle` optimised for client-side local evaluation.
        Supports ETag-based conditional requests — clients should send `If-None-Match`
        to avoid re-downloading an unchanged bundle. Cached for 60 seconds.
      operationId: getPolicyBundle
      security:
        - ApiKeyAuth: []
      parameters:
        - name: If-None-Match
          in: header
          schema:
            type: string
          description: ETag from a previous response (returns 304 if unchanged)
      responses:
        '200':
          description: Compiled policy bundle
          headers:
            ETag:
              schema:
                type: string
              description: Bundle checksum for conditional requests
            Cache-Control:
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                description: Compiled PolicyBundle (SDK-compatible format)
                properties:
                  checksum:
                    type: string
                  compiledAt:
                    type: string
                    format: date-time
                  rules:
                    type: array
                    items:
                      type: object
        '304':
          description: Not Modified (ETag matched — bundle is unchanged)
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /policy/coverage:
    post:
      tags: [Policy]
      summary: Check policy coverage for tools
      description: |
        Dry-run a list of tool names through the policy engine and return per-tool
        coverage decisions. Used to identify gaps before agent certification.
      operationId: checkPolicyCoverage
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tools]
              properties:
                tools:
                  type: array
                  minItems: 1
                  maxItems: 200
                  items:
                    type: string
                    minLength: 1
                  description: Tool names to check coverage for
            example:
              tools: ["file_read", "shell_exec", "db_query", "send_email"]
      responses:
        '200':
          description: Coverage results
          content:
            application/json:
              schema:
                type: object
                properties:
                  coveragePercent:
                    type: number
                    description: Percentage of tools with explicit policy rules
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        tool:
                          type: string
                        decision:
                          type: string
                          enum: [allow, block, monitor, require_approval]
                        ruleId:
                          type: string
                          nullable: true
                        riskScore:
                          type: integer
                        covered:
                          type: boolean
                          description: True if matched an explicit rule (not default)
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /policy/versions:
    get:
      tags: [Policy]
      summary: List policy versions
      description: |
        List all saved policy versions for the authenticated tenant (up to the most recent 50).
        Use POST /policy/revert/{version} to restore a previous version.
      operationId: listPolicyVersions
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Policy version history
          content:
            application/json:
              schema:
                type: object
                properties:
                  tenantId:
                    type: string
                  versions:
                    type: array
                    items:
                      type: object
                      properties:
                        version:
                          type: integer
                        savedAt:
                          type: string
                          format: date-time
                        ruleCount:
                          type: integer
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /policy/revert/{version}:
    post:
      tags: [Policy]
      summary: Revert policy to a previous version
      description: |
        Restore the policy to a previously saved version. The current policy is
        automatically saved before reverting (creating a new version).
      operationId: revertPolicy
      security:
        - ApiKeyAuth: []
      parameters:
        - name: version
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
          description: Version number to revert to (from GET /policy/versions)
      responses:
        '200':
          description: Policy reverted
          content:
            application/json:
              schema:
                type: object
                properties:
                  tenantId:
                    type: string
                  revertedTo:
                    type: integer
                    description: Version number restored
                  message:
                    type: string
        '400':
          description: Invalid version number
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Version not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── POLICY GIT ────────────────────────────────────────────────────────────
...